intergrations/GreenStemGlobalA1/Comms_Agents/pyproject.toml
[project]
name = "comms-agents"
version = "0.1.0"
description = "Comms & Influence Agents — MCP/MAR/Orion-ready"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2",
  "python-dotenv",
  "celery",
  "redis",
  "psycopg2-binary",
  "sqlalchemy",
  "chromadb",
  "httpx",
  "typing-extensions",
  "posthog",
  "markdown-it-py",
  "langchain>=0.1.0",
  "langgraph>=0.0.20",
  "langchain-openai>=0.0.5",
  "langchain-google-genai>=0.0.6",
  "faiss-cpu>=1.7.4",
  "tiktoken>=0.5.0",
  "transformers>=4.35.2",
  "torch>=1.1.0",
  "requests>=2.31.0",
  "pandas>=2.1.4",
  "numpy>=1.24.3",
  "google-cloud-documentai>=2.21.0",
  "google-cloud-storage>=2.10.0",
  "google-api-python-client>=2.108.0",
  "googlesearch-python>=1.2.3",
  "streamlit>=1.28.1",
  "plotly>=5.17.0",
  "scikit-learn>=1.3.2",
  "PyPDF2>=3.0.1",
  "pdfplumber>=0.10.3",
  "openpyxl>=3.1.2",
]

[project.optional-dependencies]
llama = [
  "llama-index-core",
  "llama-index-embeddings-openai",
  "llama-index-vector-stores-chroma",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

